requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
certifi==2018.1.18
chardet==3.0.4
coverage==4.4.2
decorator==4.2.1
Django==2.0.1
django-cors-headers==2.1.0
django-oauth-toolkit==1.0.0
djangorestframework==3.7.7
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.1
kombu==4.1.0
oauthlib==2.0.6
parso==0.1.1
pexpect==4.3.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
ptyprocess==0.5.2
Pygments==2.2.0
pytz==2017.3
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
tweepy==3.5.0
twython==3.6.0
urllib3==1.22
vine==1.1.4
wcwidth==0.1.7
